chore(structex): untrack repo-root clutter and relocate root tests (HEALTH-1 #8258)
Untrack (git rm --cached, files kept on disk) 8 landing-*.png,
openrouter-auth-page.png, the stale root openapi.json (canonical copy stays at
docs/api/openapi.json), and gmail_failures.json; extend .gitignore to keep them
out. Relocate root test_debate.py -> tests/debate/test_root_debate.py and
test_replay_integration.py -> tests/integration/test_replay_integration.py with
all test functions preserved (live-agent debate test skipped). File the
gmail_failures.json history decision item under docs/governance/.

Refs #8258 #8257
