--- gmail_failures.json
@@ -1,14 +1,0 @@
-{
-  "scarmani@gmail.com": {
-    "total_emails": 0,
-    "by_type": {},
-    "by_severity": {},
-    "by_sender": {},
-    "date_range": [
-      "",
-      ""
-    ],
-    "top_subjects": [],
-    "emails": []
-  }
-}
